Função de cadastro nas classes funciona no front e backend

+++ src/persistencia/Produto.java
@@ -14,7 +14,7 @@ public class Produto {
     private String categoria;
     private String observacoes;
     private int quantidade;
-    private float preco;
+    private double preco;
 
     public int getId() { return id; }
     public void setId(int id) { this.id = id; }
@@ -31,6 +31,6 @@ public class Produto {
     public int getQuantidade() { return quantidade; }
     public void setQuantidade(int quantidade) { this.quantidade = quantidade; }
 
-    public float getPreco() { return preco; }
-    public void setPreco(float preco) { this.preco = preco; }
+    public double getPreco() { return preco; }
+    public void setPreco(double preco) { this.preco = preco; }
 }
